# Backend/openapi.yaml
openapi: 3.1.0
info:
  title: SCUD Backend API
  version: "1.0.0"
  description: |
    HTTP API системы управления доступом (SCUD).

    Контракт описывает три класса участников:

    - **Мобильное приложение пользователя** (Android) — авторизуется
      session_token'ом, выпускает ключи доступа, скачивает Bloom-фильтры для
      доставки на ридер, отправляет отчёты от ридера на сервер.
    - **Администратор / интеграция** — авторизуется API-ключом
      (`X-Api-Key`), управляет пользователями, группами, ридерами, permit'ами,
      выпущенными ключами, наблюдает доставку фильтров.
    - **Public** — health-check и пара unauthenticated auth-эндпоинтов
      (login / refresh).

    Ключевые понятия:

    - **Reader** — физический контроллер двери (ESP32). Имеет 16-байтный
      `reader_id` и Ed25519 публичный ключ.
    - **Reader group** — группа ридеров; пользователь видит только ридеры
      из доступной ему `user_group → reader_group` пары.
    - **Permit** — разрешение «пользователь P может пользоваться ридером R
      в окне [valid_from, valid_until], выпускать до n_parallel ключей с
      TTL ≤ max_token_ttl_seconds».
    - **IssuedKey** — выпущенный пользователем ключ (CARD на терминологию
      ридера). Подписан сервером, содержит `permit_id`, `device_id`, окно
      `[issued_at, expires_at]`. Отозванные ключи попадают в Bloom-фильтр
      ридера.
    - **TimeGrant** — подписанный сервером talisman, разрешающий телефону
...
      `"revoked"`).
servers:
...
security:
  - SessionToken: []
  - AdminApiKey: []
paths:
...
  /api/v1/admin/readers/enroll:
    post:
      tags: [admin-readers]
      summary: Enroll нового ридера
      description: |
        Регистрирует ESP32-ридер: его 16-байтный `reader_id`, Ed25519 pubkey
        и группу. В ответе сервер возвращает свои Ed25519/X25519 pubkey'и —
        прошивка должна сохранить их (обычно через provisioning UART), чтобы
        дальше валидировать подписи серверных артефактов (filter package,
        time-grant, и т.д.).
      security:
        - AdminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/EnrollReaderRequest"}
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: "#/components/schemas/EnrollReaderResponse"}}}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "group_not_found"}
        "409": {description: "reader_id_exists"}
        "422": {description: "invalid hex / wrong byte length"}
  /api/v1/admin/readers:
    get:
      tags: [admin-readers]
      summary: Список ридеров
      description: |
        Постраничный список ридеров с базовой телеметрией: время последнего
        контакта, последняя применённая версия фильтра. Может фильтроваться
        по `group_id`.
      security:
        - AdminApiKey: []
      parameters:
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
        - {name: group_id, in: query, required: false, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/AdminReader"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
  /api/v1/admin/readers/{reader_id_hex}:
    get:
      tags: [admin-readers]
      summary: Подробности ридера
      security:
        - AdminApiKey: []
      parameters:
        - {name: reader_id_hex, in: path, required: true, schema: {$ref: "#/components/schemas/ReaderIdHex"}}
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: "#/components/schemas/AdminReaderDetail"}}}
        "400": {description: "invalid hex"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "reader_not_found"}
    patch:
      tags: [admin-readers]
      summary: Изменить ридер
      description: Меняет display_name, description, reader_group_id или is_active.
      security:
        - AdminApiKey: []
      parameters:
        - {name: reader_id_hex, in: path, required: true, schema: {$ref: "#/components/schemas/ReaderIdHex"}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PatchReaderRequest"}
      responses:
        "200": {$ref: "#/components/responses/Ok"}
        "400": {description: "invalid hex"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "reader_not_found"}

  # ===================================================================
  # ADMIN / READER GROUPS
  # ===================================================================
  /api/v1/admin/reader-groups:
    post:
      tags: [admin-reader-groups]
      summary: Создать группу ридеров
      security:
        - AdminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string}
                description: {type: string, nullable: true}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  group_id: {type: string, format: uuid}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
    get:
      tags: [admin-reader-groups]
      summary: Список групп ридеров
      security:
        - AdminApiKey: []
      parameters:
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/ReaderGroup"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
...
components:
  securitySchemes:
    SessionToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        `Authorization: Bearer <session_token>` — выдаётся `/auth/login` или
        `/auth/refresh`. Хранится на сервере в таблице sessions, валидируется
        каждый запрос (revoked_at IS NULL, expires_at > now).
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        `X-Api-Key: <plaintext>` — выдаётся `/admin/api-keys` один раз при
        создании. На сервере хранится только SHA256-хеш + `key_prefix`.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: {type: integer, minimum: 0, default: 0}
      description: Offset для пагинации.
    Limit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Ok"}
...
    AdminUnauthorized:
      description: missing_api_key / invalid_api_key / api_key_expired
  schemas:
    Ok:
      type: object
      properties:
        ok: {type: boolean, example: true}
...
    ReaderIdHex:
      type: string
      pattern: "^[0-9a-fA-F]{32}$"
      description: 16-байтный reader_id в hex (32 символа).
      example: "a1b2c3d4e5f60718293a4b5c6d7e8f90"

    # ----- AUTH -----
...
    AdminReader:
      type: object
      properties:
        reader_id: {type: string, description: "hex"}
        display_name: {type: string}
        description: {type: string, nullable: true}
        reader_group_id: {type: string, format: uuid}
        is_active: {type: boolean}
        enrolled_at: {type: string, format: date-time}
        last_contact_at: {type: string, format: date-time, nullable: true}
        last_applied_filter_version: {type: integer}
    AdminReaderDetail:
      allOf:
        - {$ref: "#/components/schemas/AdminReader"}
        - type: object
          properties:
            last_known_time: {type: string, format: date-time, nullable: true}
            open_delivery_tasks_count: {type: integer}
    PatchReaderRequest:
      type: object
      properties:
        display_name: {type: string}
        description: {type: string, nullable: true}
        reader_group_id: {type: string, format: uuid}
        is_active: {type: boolean}
    EnrollReaderRequest:
      type: object
      required: [reader_id, reader_pubkey, reader_group_id, display_name]
      properties:
        reader_id: {$ref: "#/components/schemas/ReaderIdHex"}
        reader_pubkey: {type: string, description: "base64, 32 bytes Ed25519"}
        reader_group_id: {type: string, format: uuid}
        display_name: {type: string}
        description: {type: string, nullable: true}
    EnrollReaderResponse:
      type: object
      properties:
        reader_id: {type: string}
        server_ed25519_pubkey: {type: string, description: "base64, 32 bytes"}
        server_x25519_pubkey: {type: string, description: "base64, 32 bytes"}
        reader_group_id: {type: string, format: uuid}
    ReaderGroup:
      type: object
      properties:
        group_id: {type: string, format: uuid}
        name: {type: string}
        description: {type: string, nullable: true}
        created_at: {type: string, format: date-time}

    # ----- USERS (ADMIN) -----
...